crs/src/requirements.txt
requests
pyyaml
litellm
jq
GitPython